Refactor: Hide saved login users behind UserRepository
- Added `getSavedUsers`, `saveSavedUser`, and `resetGuestAsMember` methods to `UserRepository`.
- Implemented duplicate-user handling rules in `UserRepositoryImpl`.
- Refactored `LoginViewModel` to delegate saved-user persistence to `UserRepository` instead of accessing `SharedPrefManager` directly.
- Updated `LoginViewModelTest` to mock `UserRepository` accordingly.
