+++ app/src/main/java/org/ole/planet/myplanet/repository/UserRepository.kt
@@ -7,8 +7,12 @@ import org.ole.planet.myplanet.model.HealthRecord
 import org.ole.planet.myplanet.model.RealmAchievement
 import org.ole.planet.myplanet.model.RealmMyHealth
 import org.ole.planet.myplanet.model.RealmUser
+import org.ole.planet.myplanet.model.User
 
 interface UserRepository {
+    suspend fun getSavedUsers(): List<User>
+    suspend fun saveSavedUser(name: String?, encryptedPassword: String?, source: String, userProfile: String?, userName: String?)
+    suspend fun resetGuestAsMember(username: String?)
     suspend fun getHealthProfile(userId: String): RealmMyHealth?
     suspend fun updateUserHealthProfile(userId: String, userData: Map<String, Any?>)
