+++ app/src/main/java/org/ole/planet/myplanet/ui/sync/LoginViewModel.kt
@@ -12,13 +12,13 @@ import org.ole.planet.myplanet.model.RealmMyTeam
 import org.ole.planet.myplanet.model.RealmUser
 import org.ole.planet.myplanet.model.User
 import org.ole.planet.myplanet.repository.TeamsRepository
-import org.ole.planet.myplanet.services.SharedPrefManager
+import org.ole.planet.myplanet.repository.UserRepository
 import org.ole.planet.myplanet.utils.DispatcherProvider
 
 @HiltViewModel
 class LoginViewModel @Inject constructor(
     private val teamsRepository: TeamsRepository,
-    private val sharedPrefManager: SharedPrefManager,
+    private val userRepository: UserRepository,
     private val dispatcherProvider: DispatcherProvider
 ) : ViewModel() {
 
@@ -59,70 +59,21 @@ class LoginViewModel @Inject constructor(
 
     fun loadSavedUsers() {
         viewModelScope.launch(dispatcherProvider.io) {
-            _savedUsers.value = sharedPrefManager.getSavedUsers()
+            _savedUsers.value = userRepository.getSavedUsers()
         }
     }
 
     fun saveUsers(name: String?, encryptedPassword: String?, source: String, userProfile: String?, userName: String?) {
         viewModelScope.launch(dispatcherProvider.io) {
-            val existingUsers: MutableList<User> = ArrayList(sharedPrefManager.getSavedUsers())
-            if (source == "guest") {
-                val newUser = User("", name, encryptedPassword, "", "guest")
-                var newUserIndex = -1
-                for (i in existingUsers.indices) {
-                    if (existingUsers[i].name == newUser.name?.trim { it <= ' ' }) {
-                        newUserIndex = i
-                        break
-                    }
-                }
-                if (newUserIndex != -1) {
-                    existingUsers[newUserIndex] = newUser
-                } else {
-                    existingUsers.add(newUser)
-                }
-                sharedPrefManager.setSavedUsers(existingUsers)
-            } else if (source == "member") {
-
-                val newUser = User(userName, name, encryptedPassword, userProfile, "member")
-                var newUserIndex = -1
-                for (i in existingUsers.indices) {
-                    if (existingUsers[i].fullName == newUser.fullName?.trim { it <= ' ' }) {
-                        newUserIndex = i
-                        break
-                    }
-                }
-                if (newUserIndex != -1) {
-                    existingUsers[newUserIndex] = newUser
-                } else {
-                    existingUsers.add(newUser)
-                }
-                sharedPrefManager.setSavedUsers(existingUsers)
-            }
+            userRepository.saveSavedUser(name, encryptedPassword, source, userProfile, userName)
             loadSavedUsers()
         }
     }
 
     fun resetGuestAsMember(username: String?) {
         viewModelScope.launch(dispatcherProvider.io) {
-            val existingUsers = sharedPrefManager.getSavedUsers().toMutableList()
-            var newUserExists = false
-            for ((_, name) in existingUsers) {
-                if (name == username) {
-                    newUserExists = true
-                    break
-                }
-            }
-            if (newUserExists) {
-                val iterator = existingUsers.iterator()
-                while (iterator.hasNext()) {
-                    val (_, name) = iterator.next()
-                    if (name == username) {
-                        iterator.remove()
-                    }
-                }
-                sharedPrefManager.setSavedUsers(existingUsers)
-                loadSavedUsers()
-            }
+            userRepository.resetGuestAsMember(username)
+            loadSavedUsers()
         }
     }
 }
